+++ src/rubysys/mod.rs
@@ -8,6 +8,7 @@ pub mod fixnum;
 pub mod gc;
 pub mod float;
 pub mod hash;
+pub mod onig;
 pub mod rproc;
 pub mod string;
 pub mod symbol;